+++ src/backend/TrafficCourts/Workflow.Service/Services/EmailSenderService.cs
@@ -4,6 +4,7 @@ using TrafficCourts.Workflow.Service.Configuration;
 using TrafficCourts.Common.OpenAPIs.OracleDataApi.v1_0;
 using MimeKit;
 using MimeKit.Text;
+using TrafficCourts.Common.Features.Mail.Model;
 
 namespace TrafficCourts.Workflow.Service.Services
 {
@@ -205,6 +206,25 @@ namespace TrafficCourts.Workflow.Service.Services
             }
 
             return false;
+        }
+
+        public SendEmail ToVerificationSendEmail(Dispute dispute, string host)
+        {
+            SendEmail sendEmail = new();
+            // Send email message to the submitter's entered email
+            var template = MailTemplateCollection.DefaultMailTemplateCollection.FirstOrDefault(t => t.TemplateName == "VerificationEmailTemplate");
+            if (template is not null)
+            {
+                sendEmail.From = template.Sender;
+                sendEmail.To.Add(dispute.EmailAddress);
+                sendEmail.Subject = template.SubjectTemplate.Replace("<ticketid>", dispute.TicketNumber);
+                sendEmail.PlainTextContent = template.PlainContentTemplate?.Replace("<ticketid>", dispute.TicketNumber);
+                sendEmail.HtmlContent = template.HtmlContentTemplate?.Replace("<ticketid>", dispute.TicketNumber);
+                sendEmail.HtmlContent = sendEmail.HtmlContent?.Replace("<emailverificationtoken>", dispute.EmailVerificationToken);
+                sendEmail.HtmlContent = sendEmail.HtmlContent?.Replace("<baseref>", host);
+                sendEmail.TicketNumber = dispute.TicketNumber;
+            }
+            return sendEmail;
         }
 
         /// <summary>
