+++ src/backend/TrafficCourts/Workflow.Service/Services/IEmailSenderService.cs
@@ -1,9 +1,12 @@
 ﻿using TrafficCourts.Messaging.MessageContracts;
+using TrafficCourts.Common.OpenAPIs.OracleDataApi.v1_0;
 
 namespace TrafficCourts.Workflow.Service.Services
 {
     public interface IEmailSenderService
     {
         Task SendEmailAsync(SendEmail emailMessage, CancellationToken cancellationToken);
+        SendEmail ToVerificationSendEmail(Dispute dispute, string host);
+
     }
 }
