+++ src/backend/TrafficCourts/Workflow.Service/TrafficCourts.Workflow.Service.csproj
@@ -24,7 +24,6 @@
   <ItemGroup>
     <ProjectReference Include="..\Common\TrafficCourts.Common.csproj" />
     <ProjectReference Include="..\Messaging\TrafficCourts.Messaging.csproj" />
-    <ProjectReference Include="..\Staff.Service\TrafficCourts.Staff.Service.csproj" />
     <ProjectReference Include="..\TrafficCourts.Arc.Dispute.Client\TrafficCourts.Arc.Dispute.Client.csproj" />
   </ItemGroup>
 </Project>
